+++ tests/Public/Save-EvergreenApp.Tests.ps1
@@ -3,8 +3,8 @@
         Public Pester function tests.
 #>
 [OutputType()]
-[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments", "", Justification="This OK for the tests files.")]
-[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSAvoidUsingWriteHost", "", Justification="Outputs to log host.")]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSUseDeclaredVarsMoreThanAssignments", "", Justification = "This OK for the tests files.")]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute("PSAvoidUsingWriteHost", "", Justification = "Outputs to log host.")]
 param ()
 
 BeforeDiscovery {
@@ -32,14 +32,32 @@ Describe -Tag "Save" -Name "Save-EvergreenApp" -ForEach $Installers {
 
     # Test that Save-EvergreenApp accepts the object and saves the file
     Context "Validate Save-EvergreenApp works with <installer.Architecture>." {
-        It "Save-EvergreenApp should not Throw" {
-            { $File = $installer | Save-EvergreenApp -Path $Path } | Should -Not -Throw
+        It "Save-EvergreenApp should not Throw with Path" {
+            $params = @{
+                InputObject = $installer
+                Path        = $Path
+                UserAgent   = [Microsoft.PowerShell.Commands.PSUserAgent]::Firefox
+                Force       = $true
+                NoProgress  = $true
+            }
+            { Save-EvergreenApp @params } | Should -Not -Throw
         }
 
         # Test that the file downloaded into the path: "$Path/Stable/Enterprise/<version>/x64/MicrosoftEdgeEnterpriseX64.msi"
         It "Should save in the right path" {
             $File = [System.IO.Path]::Combine($Path, $installer.Channel, $installer.Release, $installer.Version, $installer.Architecture, $(Split-Path -Path $installer.URI -Leaf))
             Test-Path -Path $File -PathType "Leaf" | Should -Be $True
+        }
+
+        It "Save-EvergreenApp should not Throw with CustomPath" {
+            $params = @{
+                InputObject = $installer
+                CustomPath  = $Path
+                UserAgent   = [Microsoft.PowerShell.Commands.PSUserAgent]::Firefox
+                Force       = $true
+                NoProgress  = $true
+            }
+            { Save-EvergreenApp @params } | Should -Not -Throw
         }
     }
 }