+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Change log
 
-## VERSION
+## 2212.711
 
 * Adds `MicrosoftVdot`
 * Adds parameter `-UserAgent` to `Test-EvergreenApp`